requirements.txt
# FastAPI y servidor ASGI
fastapi==0.100.0
uvicorn==0.23.1

# Conexión a Snowflake
snowflake-connector-python==3.0.5

# Machine Learning
scikit-learn==1.3.0
joblib==1.3.2
numpy==1.24.4
pandas==2.1.1

# Visualización y explicabilidad
shap==0.42.1
matplotlib==3.8.1

# Manejo de variables de entorno
python-dotenv==1.0.0

# Opcional (para pruebas)
pytest==7.5.0
httpx==0.24.1
